appointments added

+++ pkg/model/doctor.go
@@ -2,7 +2,7 @@ package model
 
 //Doctor data
 type Doctor struct {
-	ID       int64  `json:"id"`
+	ID       uint   `json:"id"`
 	FullName string `json:"full_name"`
 	Gain     int64  `json:"gain"`
 	Phone    string `json:"phone"`
